_requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
ExifRead==2.3.2
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.3
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-cloud-vision==2.3.1
google-pasta==0.2.0
googleapis-common-protos==1.53.0
#grpcio==1.32.0
h5py==2.10.0
idna==2.10
image-quality==1.2.7
imageio==2.9.0
joblib==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libsvm==3.23.0.4
Markdown==3.3.4
matplotlib==3.4.1
networkx==2.5.1
numpy==1.19.5
oauthlib==3.1.0
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
Pillow==8.2.0
proto-plus==1.18.1
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybrisque==1.0
pycparser==2.20
pyheif==0.5.1
pyparsing==2.4.7
pyping==0.0.6
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
tensorflow-serving-api
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
